requirements-server.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
easyocr
pillow
pydantic>=2.0.0
